test(voxcpm): split unit and plugin test modules
Each test module must declare exactly one category marker per AGENTS.md.
Move the live vLLM-Omni integration test into test_plugin_voxcpm_tts.py.
